refactor(wip)!: code structure
